% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowgill_setup.R
\name{cowgill_setup}
\alias{cowgill_setup}
\title{Setup an R install with my most-used packages and key settings}
\usage{
cowgill_setup()
}
\description{
Installs CRAN and GitHub packages, installs rio formats, rsthemes, and
fonts.
}
\examples{
\dontrun{
cowgill_setup()
}
}
